+++ src/settings/MappingProfiles/index.js
@@ -1,2 +1,3 @@
 export * from './MappingProfiles';
 export * from './ViewMappingProfile';
+export * from './MappingProfilesForm';